backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.34.0
pydantic==2.8.2
pydantic-core==2.20.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
aiofiles==24.1.0
gitpython==3.1.43
radon==6.0.1
pygments==2.19.1
networkx==3.4.2
matplotlib==3.10.0
python-dotenv==1.0.1
alembic==1.14.0
psycopg2-binary==2.9.10
requests==2.32.3
celery==5.4.0
redis==5.2.1
email-validator==2.2.0
google-auth==2.38.0
razorpay==1.4.2
